Changing isPermalink back to "false" for the default RSS story template.
Reason:

The URL in there is no valid permalink and could cause non-working links
to the postings because of the missing flavour suffix in dynamic as well
as static setups. $default_flavour only applies to index pages, e.g.
start page or per year/month/day pages.

Thanks to Lilo von Hanffstengel for pointing this out and insisting on
me being wrong. :-)