Fix static rendering with entries subroutine
When we have "article.txt" in a directory named "mytheme", we got a file
"mythemearticle.html" generated at the top level of static directory.
What is expected is a file "article.html" generated in the directory
"mytheme" which itself is contained by the static directory.

The bug was introduced by commit 1d9899f : a test done on $1 being empty
or not was not transposed.
I became aware of the bug only today, as I routinely use the
"entries_cache" plugin as a substitute of the default "entries"
subroutine.